requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
gunicorn==21.2.0
requests==2.32.5
python-dotenv==1.2.1

fastapi==0.121.2
uvicorn==0.38.0
pydantic==2.12.4
starlette==0.49.3
typing_extensions==4.15.0
typing-inspection==0.4.2
h11==0.16.0
anyio==4.11.0
python-multipart==0.0.20
annotated-types==0.7.0
pydantic_core==2.41.5

openai==1.51.0
